feat(backend & frontend): \ -after successfully oauth and after generate token, set token to store inside cookie as key name AUTH_TOKEN\ -add userId as claim in jwt token\\ -add endpoints to create workspaces and boards , along with the endpoints to retrieve existing workspaces and boards\

+++ authservice/src/main/java/com/server/authservice/repository/UserRepo.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 
 // Repository used to interact with the User table in the database
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepo extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
